Check for png/svg before copying

+++ production.sh
@@ -6,6 +6,11 @@ cp public/build/manifest.webmanifest public
 cp public/build/manifest.webmanifest public/manifest.json
 
 # All static assets must be listed here to show up
-cp public/build/*.png public
-cp public/build/*.svg public
+if [ -f public/build/*.png]
+    cp public/build/*.png public
+fi
+
+if [ -f public/build/*.svg]
+    cp public/build/*.svg public
+fi
 # cp public/build/*.ico public
